Connect sql window with qep window
Aside from printing the tree correctly, the qep tree will also show
relations show coloured in light gray. The qep tree window will only
show the boundaries of the entire tree and scrolling can be used to view
parts of the tree.

Connected the user enquiry to get the tree.